set buffer size from command line

+++ device/defs.hpp
@@ -7,7 +7,8 @@
 #define GLOBAL_HEADER_SIZE sizeof(struct global_header)
 #define CHUNK_SIZE sizeof(struct chunk)
 
-extern int BUFFER_SIZE;
+// defined in main_cpu.cpp
+extern size_t BUFFER_SIZE;
 extern int CHUNKS_PER_BUFFER;
 
 const int BUFFER_COUNT = 2;
